Modified Format, Edit semi-passive Binary Engine
Changed the __init__ for the VMMEngine. I now see the __init__ as a form of "purchasing a chip from a vendor", where the parameters is just specifying to the vendor what type of chip you want to purchase.

This includes the memristor type, crossbar dimensions, ADC classes... etc. We also initialize a crossbar in here, because it's the same as if we are buying a crossbar chip.